Move unit tests into its own directory
