+++ .github/workflows/ci-cd.yml
@@ -38,15 +38,15 @@ jobs:
         run: npm ci --unsafe-perm
         working-directory: init
 
+      - name: Run unit tests
+        run: npm run test:unit -- --coverage
+        working-directory: init
+
       - name: Docker Compose version
         run: docker-compose version --short
 
       - name: Start services
         run: docker-compose up --quiet-pull -d
-
-      - name: Run tests
-        run: npm test -- --coverage
-        working-directory: init
 
       - name: Show logs
         run: docker-compose logs --tail all
